Move async validation to the same class with the blocking one
