+++ Px.Utils.UnitTests/SyntaxValidationTests/StreamSyntaxValidationAsyncTests.cs
@@ -20,7 +20,7 @@ namespace PxUtils.UnitTests.SyntaxValidationTests
             stream.Seek(0, SeekOrigin.Begin);
 
             // Act
-            SyntaxValidationResult result = await SyntaxValidationAsync.ValidatePxFileMetadataSyntaxAsync(stream, filename);
+            SyntaxValidationResult result = await SyntaxValidation.ValidatePxFileMetadataSyntaxAsync(stream, filename);
 
             Assert.AreEqual(8, result.StructuredEntries.Count);
             Assert.AreEqual(0, result.Report.FeedbackItems?.Count);
@@ -35,7 +35,7 @@ namespace PxUtils.UnitTests.SyntaxValidationTests
             stream.Seek(0, SeekOrigin.Begin);
 
             // Act
-            SyntaxValidationResult result = await SyntaxValidationAsync.ValidatePxFileMetadataSyntaxAsync(stream, filename);
+            SyntaxValidationResult result = await SyntaxValidation.ValidatePxFileMetadataSyntaxAsync(stream, filename);
 
             Assert.AreEqual(10, result.StructuredEntries.Count);
             Assert.AreEqual("YES", result.StructuredEntries[8].Value);
@@ -56,7 +56,7 @@ namespace PxUtils.UnitTests.SyntaxValidationTests
             stream.Seek(0, SeekOrigin.Begin);
 
             // Act
-            SyntaxValidationResult result = await SyntaxValidationAsync.ValidatePxFileMetadataSyntaxAsync(stream, filename);
+            SyntaxValidationResult result = await SyntaxValidation.ValidatePxFileMetadataSyntaxAsync(stream, filename);
 
             Assert.AreEqual(1, result.Report.FeedbackItems?.Count);
         }
